Fix the validation and updated test cases

+++ src/Apha.FPS/Apha.FPS.Application/Services/ProjectService.cs
@@ -1,6 +1,7 @@
 ﻿using Apha.FPS.Application.Dtos;
 using Apha.FPS.Application.Interfaces;
 using Apha.FPS.Application.Pagination;
+using Apha.FPS.Application.Validation;
 using Apha.FPS.Core.Entities;
 using Apha.FPS.Core.Interfaces;
 using Apha.FPS.Core.Pagination;
@@ -68,8 +69,19 @@ namespace Apha.FPS.Application.Services
 
         public async Task<bool> DeleteProjectAsync(string parentProject)
         {
+            var hasAssociations = await _projectRepository.HasAssociatedJobCodesAsync(parentProject);
+            if (hasAssociations)
+            {
+                throw new BusinessValidationErrorException(
+                [
+                    new BusinessValidationError(
+                        $"Project '{parentProject}' is associated with existing job codes and cannot be deleted.",
+                        "PROJECT_HAS_ASSOCIATIONS")
+                ]);
+            }
+
             return await _projectRepository.DeleteProjectAsync(parentProject);
-        }   
+        }
 
         public async Task<PaginatedResult<ProjectDto>> GetProjectsByProgramAsync(QueryParameters<string> query, string programNo)
         {
